pyproject.toml
[project]
name = "morgue-rag"
version = "0.1.0"
description = "A CLI tool to explore Graph Based Retrieval-Augmented Generation (GraphRAG)"
authors = [{ name = "LukV", email = "luk.verhelst@vlaanderen.be" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"

dependencies = [
    "duckdb>=0.10.1",
    "rdflib>=7.1.4",
    "typer>=0.15.4",
    "llama-index>=0.10.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-embeddings-huggingface>=0.3.0",
    "llama-index-llms-ollama>=0.3.0",
    "llama-index-vector-stores-duckdb>=0.3.1",
    "openai>=1.0.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sentence-transformers>=2.2.0",
]

[dependency-groups]
dev = [
    "commitizen>=4.8.2",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.11",
]

[project.scripts]
morgue-rag = "cli:app"

[tool.uv]
package = true

[[tool.mypy.overrides]]
module = ["pandas.*"]
ignore_missing_imports = true
